copier: fix for files with / in their names
If a file in the source pot had a slash in its name, everything before the
slash would be lost when rsyncing the file to the destination pot, for the
same reason that this happens when typing "cp a/b/c d/". Fortunately, it turns
out that rsync has some nice features built in that make it easy to get The
Right Thing to happen and preserve the sub-pot directory structure.

The feature to solve this was introduced in rsync 2.6.7; if any of our
deployed systems use an older version, we'll need to take another approach.