Port from MySQL to PostgreSQL
This is lightly tested; mistakes may linger.

One of the changes is to modify some of the record-creating RPC functions to
not return the "id" column the newly-created database records. The proximate
cause was that Postgres returns IDs in a different way than MySQL, but it's a
correctness issue: the SQL database is an Librarian server implementation
detail, and it's not right to leak its contents to RPC clients.

Also, the delete_when_done code in `copier` was apparently untested: it had a
wrong column name and called a function that didn't exist. It also doesn't
look like it will delete successfully if it's trying to delete a MIRIAD
dataset since those are directories.