Remove M&C-related stub files.
These were apparently stubbed for testing interactions between the Librarian
and the telescope Monitor & Control subsystem, but implementation of M&C
has been moved to a new repository now:

https://github.com/HERA-Team/Monitor_and_Control

These files are only used by commented-out code in the current version of the
Librarian, so let's just clean them out. If we really find ourselves missing
the old code we can always dig it out of the Git history.