Implement inter-Librarian copying.
We now have an API call that will make a Librarian go to one of its stores
and launch a copy process that scp's to a store in a different Librarian.

Loose ends: we don't monitor the success of the copy whatsoever! We in fact
completely lose the output of the copy process. And there's no rate-limiting.
Etc.

What I should do is have the Librarian server run a long-running SSH process
to the local store, which in turn starts an SCP to the remote store.