Drastically clean up Librarian copying and add a command-line tool for it.
The key thing is that we don't need to give the Librarian a local store name
and path -- file names are unique, and the Librarian knows about all file
instances, so given a name it can find an instance, and any instance will do.
This makes the launch_librarian_copy.py script quite simple.