Add an 'ObservingSession' concept to the Librarian.
This allows us to group many Observations into a session that is contiguous
in some sense -- ideally a continuous set of observations taken during one night
of observing. This concept is needed to be able to correctly populate the RTP with
data, since it needs to know the "neighbors" of every observation in order to
be able to do its work. But this should be a more broadly useful concept since it
is quite reasonable to consider data on a per-night granularity in a lot of
circumstances.

We haven't implemented any user interface or inter-Librarian for sessions at this
point, though.