server: implement describe_session_without_event API
This is the latest iteration of the API needed to allow the RTP system to
query the Librarian for new data to process.

It turns out that (1) the RTP system needs to be told about whole batches of
files at a time so that it can do its "neighbor" linking and (2) it doesn't
drop information about files once they've been processed. Therefore the way to
do this is for the Librarian to tell the RTP about all of the files in a
single observing run in one big batch. This API attempts to do so.