Massive rework of file uploading.
Three big changes here. First, we now upload file metadata (basically: File,
Observation, and ObservingSession records, serialized to JSON) before we
actually upload the data. The destination Librarian ingests those records and
so doesn't need to infer the metadata from the bare files alone, which it
isn't able to completely do in general.

To make this possible, the upload-completion code gained two modes: in one
mode, we assume that the above has happened and that nothing needs to be
inferred. In that case, we require that the needed records exist and verify
that the file matches them. In the other, we *do* do the inference, which will
only work if the file has a nicely-formatted name. In the future we may also
require that it be a valid UV dataset so that we can learn its duration, etc.

Second, uploads now go into a temporary "staging" directory with a uniquely
generated temporary name. This is just a better design that what I had before.

Third, inter-Librarian uploads are now conducted in a background thread inside
the Librarian server. This just needed to happen.