scripts/add_obs_librarian.py: tidy up and straighten out
This operated in a bit of an inefficient way where we contacted the Librarian,
which then went back and SSH'ed into us to gather the information about the
files we were just telling it about. Change the design to gather all of that
information and ship it off to the Librarian in one go.