Take some steps to avoid clobbers on upload.
The Librarian takes care not to clobber file instances of which it's aware,
but testing suggests that it's quite plausible that there will be files within
the Librarian storage directory of which it's not aware. So whenever we move
or create non-staging files, we need to be careful. To begin with, wire up the
"upload" completion logic to be more careful.