server/librarian_server/observation.py: fix initial file listing in describe_session_without_event
This feature wasn't working when the RTP actually did anything. The first step
in the function is to identify files that are assigned to a session but are
not marked as having been ingested into the RTP. However, this search did not
filter by file "source". Since the RTP adds new files to a session in the
course of processing it, that meant that the next time the Librarian was asked
to find data to process, there would be additional files from that session
matching that filter. However, the next step, where information about files
was actually gathered, *would* properly filter by source, so RTP would be sent
the same list of files it already knew about.

This was just a mistake: all file queries in this function should filter by
source, not just the final one.