server/librarian_server/observation.py: fix the fix
In the previous commit I got too cocky and didn't notice the part where I
added the extra source filter to the wrong query subexpression. THIS time it
will obviously work and I REALLY don't need to test the commit. Totally.

(I would test it but I'm on a plane so I can't log in to the Karoo. Or I
probably can but it'd cost me 8 Euros. Technology is so lame.)