Add the FileInstance.deletion_policy column.
We want to be able to delete file instances to save space on the Karoo. But,
we want some safety interlocks so that we can't accidentally blow away the
entire science data archive. The new "deletion_policy" attribute of file
instances will provide that interlock, because any instances that are to be
deleted will have to be explicitly marked as such.