Change offloading to go through the standard deletion framework.
Rather than deleting files right in the offloader task, we mark them for
deletion. The new search and deletion filtering tools that I've implemented
then make it feasible to issue a deletion command that targets the offloaded
files. The command line interface to the offloader helpfully prints out the
needed command.