Add "known staging directory" support for expedited Karoo=>NRAO uploads
Some HERA-19 data have already been copied to NRAO but were just copied
over outside of the Librarian infrastructure. We'd like to ingest them
into the new Librarian with all their metadata while leveraging the
fact that all the data have been copied.

My strategy for this is to add a "known staging" option when uploading that
lets the caller pretend that a staging directory has already been created on a
target host. This way the rsync command still runs but hopefully executes
extremely quickly if all the data are in fact available.