server: offer two views of the observing sessions
We haven't had a way to get an overview of every single piece of data on the
server. The natural way to do that is at the outermost level, with a list of
all observing sessions, but it's been very slow to compute the detailed
information for each session. So, we now add a lighter version that's much
quicker to compute, and wire up a page that shows you the basic info for every
single session. There's then a separate page that gives you detailed info for
recent sessions.