server: prep for letting people run different kinds of searches
Searches for individual files are fine, but it is very reasonable to want to
search for particular observations or observing sessions. In this commit we
start laying the groundwork for this by renaming things to be more
file-specific.

The /search API endpoint and librarian_server.search module keep their names,
since I think it will be helpful to centralize various parts of the code
across all potential search types.