Make it possible to create files with null obsids.
This functionality will support "maintenance" files that are not tied to
particular observations, such as weekly database backups. These files will not
appear in most searches since most things end up selecting on the obsid on
some way or another.

Creating such a file requires a special flag to `upload_to_librarian.py`, so
that one has to explicitly intend to create a file without an obsid. This
defends against some kind of SNAFU making it so that observation-related files
start getting ingested without obsids.